+++ tests/test_streaming.py
@@ -1,7 +1,6 @@
 from __future__ import annotations as _annotations
 
 import json
-import re
 from collections.abc import AsyncIterator
 from datetime import timezone
 
@@ -9,7 +8,7 @@ import pytest
 from inline_snapshot import snapshot
 from pydantic import BaseModel
 
-from pydantic_ai import Agent, UnexpectedModelBehavior, UsageLimitExceeded, UserError
+from pydantic_ai import Agent, UnexpectedModelBehavior, UserError
 from pydantic_ai.messages import (
     ArgsDict,
     ArgsJson,
@@ -24,7 +23,6 @@ from pydantic_ai.messages import (
 from pydantic_ai.models.function import AgentInfo, DeltaToolCall, DeltaToolCalls, FunctionModel
 from pydantic_ai.models.test import TestModel
 from pydantic_ai.result import Usage
-from pydantic_ai.settings import UsageLimits
 
 from .conftest import IsNow
 
@@ -579,43 +577,3 @@ async def test_early_strategy_does_not_apply_to_tool_calls_without_final_tool():
             ),
         ]
     )
-
-
-async def test_streamed_text_limits():
-    m = TestModel()
-
-    test_agent = Agent(m)
-    assert test_agent.name is None
-
-    @test_agent.tool_plain
-    async def ret_a(x: str) -> str:
-        return f'{x}-apple'
-
-    async with test_agent.run_stream('Hello', usage_limits=UsageLimits(response_tokens_limit=10)) as result:
-        assert test_agent.name == 'test_agent'
-        assert not result.is_structured
-        assert not result.is_complete
-        assert result.all_messages() == snapshot(
-            [
-                ModelRequest(parts=[UserPromptPart(content='Hello', timestamp=IsNow(tz=timezone.utc))]),
-                ModelResponse(
-                    parts=[ToolCallPart(tool_name='ret_a', args=ArgsDict(args_dict={'x': 'a'}))],
-                    timestamp=IsNow(tz=timezone.utc),
-                ),
-                ModelRequest(
-                    parts=[ToolReturnPart(tool_name='ret_a', content='a-apple', timestamp=IsNow(tz=timezone.utc))]
-                ),
-            ]
-        )
-        assert result.usage() == snapshot(
-            Usage(
-                requests=2,
-                request_tokens=103,
-                response_tokens=5,
-                total_tokens=108,
-            )
-        )
-        with pytest.raises(
-            UsageLimitExceeded, match=re.escape('Exceeded the response_tokens_limit of 10 (response_tokens=11)')
-        ):
-            await result.get_data()
